+++ packages/pob-dependencies/lib/index.d.ts
@@ -1,17 +1,17 @@
 
 interface Dependencies {
-  '@babel/core': '7.10.3';
-  '@babel/preset-env': '7.10.3';
-  '@babel/preset-react': '7.10.1';
-  '@babel/runtime': '7.10.3';
+  '@babel/core': '7.10.5';
+  '@babel/preset-env': '7.10.4';
+  '@babel/preset-react': '7.10.4';
+  '@babel/runtime': '7.10.5';
   '@pob/commitlint-config': '^1.1.2';
-  '@pob/eslint-config': '38.0.0';
-  '@pob/eslint-config-node': '38.0.0';
-  '@pob/eslint-config-typescript': '38.1.0';
-  '@pob/eslint-config-typescript-node': '38.0.2';
-  '@pob/eslint-config-typescript-react': '38.0.0';
+  '@pob/eslint-config': '38.2.0';
+  '@pob/eslint-config-node': '38.2.0';
+  '@pob/eslint-config-typescript': '38.2.0';
+  '@pob/eslint-config-typescript-node': '38.2.0';
+  '@pob/eslint-config-typescript-react': '38.2.0';
   '@pob/pretty-pkg': '^1.2.1';
-  '@pob/root': '^2.2.1';
+  '@pob/root': '^2.2.2';
   '@types/jest': '25.2.3';
   '@typescript-eslint/eslint-plugin': '2.34.0';
   '@typescript-eslint/parser': '2.34.0';
@@ -25,22 +25,22 @@ interface Dependencies {
   'eslint-plugin-node': '11.1.0';
   'eslint-plugin-prefer-class-properties': '1.0.0';
   'eslint-plugin-prettier': '3.1.3';
-  'eslint-plugin-react': '7.20.0';
+  'eslint-plugin-react': '7.20.3';
   'eslint-plugin-react-hooks': '2.5.1';
   'eslint-plugin-unicorn': '18.0.1';
   'husky': '4.2.5';
   'jest': '26.1.0';
   'jest-junit-reporter': '1.1.0';
   'lerna': '3.22.1';
-  'pob-babel': '^24.12.1';
-  'pob-lcov-reporter': '^2.3.0';
-  'pob-release': '^5.6.1';
+  'pob-babel': '^24.13.0';
+  'pob-lcov-reporter': '^2.4.0';
+  'pob-release': '^5.6.2';
   'prettier': '2.0.5';
   'repository-check-dirty': '^1.2.2';
   'semver': '7.3.2';
   'typedoc': '0.16.7';
-  'typedoc-plugin-lerna-packages': '^0.3.0';
-  'typescript': '3.9.5';
+  'typedoc-plugin-lerna-packages': '0.3.0';
+  'typescript': '3.9.6';
 }
 
 export = dependencies;